% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict_groups.R
\name{planarsviz_conflict_tree}
\alias{planarsviz_conflict_tree}
\title{One conflict-groups tree}
\usage{
planarsviz_conflict_tree(
  newick,
  group_spans,
  strengths,
  alphaval,
  colour = "black"
)
}
\arguments{
\item{newick}{Newick string.}

\item{group_spans}{Character vector of `"left-right"` spans in groupOTU order.}

\item{strengths}{Thickness for each span, same order.}

\item{alphaval}{Opacity of this tree.}

\item{colour}{Line colour.}
}
\value{
A ggtree plot (no tip labels).
}
\description{
One conflict-groups tree
}
